Update README.md
Devart Excel Addins v2.10.1404.0

Devart Excel Add-ins allow you to work with database and cloud data in Microsoft Excel as with usual Excel spreadsheets. With Devart Excel Add-ins you can get exactly the data you need with visual Query Builder or with SQL and refresh the queried external data in a workbook any time with a single click. External data can be edited as you usually do it in Excel, and then saved back to the data source.

[Download Devart Excel Addins v2.10.1404.0](https://developer.team/misc-development/34956-devart-excel-addins-v21014040-crack.html)